requirements.txt
mediapipe==0.8.7.2
scikit-learn
protobuf<=3.20
pyyaml
pandas
joblib

opencv-python
numpy
